# [SPARK-26205][SQL] Optimize In for bytes, shorts, ints