## What changes were proposed in this pull request?

This PR optimizes `In` expressions for byte, short, integer types. It is a follow-up on PR #21442 from @dbtsai.  

Currently, `In` expressions are compiled into a sequence of if-else statement, which results in O(n) time complexity. `InSet` is an optimized version of `In`, which is supposed to improve the performance if the number of elements is big enough. However, `InSet` actually degrades the performance in many cases due to various reasons (`InSet` will be handled in [SPARK-26203](https://issues.apache.org/jira/browse/SPARK-26203) and [SPARK-26204](https://issues.apache.org/jira/browse/SPARK-26204)).

The main idea of this PR is to make use of `tableswitch` and `lookupswitch` bytecode instructions to speed up `In` expressions. In short, we can improve our time complexity from O(n) to O(1) or at least O(log n) by using Java `switch` statements. We will have O(1) time complexity if our case values are compact and `tableswitch` can be used. Otherwise, `lookupswitch` will give us O(log n). See [here](https://docs.oracle.com/javase/specs/jvms/se7/html/jvms-3.html#jvms-3.10) and [here](https://stackoverflow.com/questions/10287700/difference-between-jvms-lookupswitch-and-tableswitch) for more information.

An important benefit of the proposed approach is that we do not have to pay an extra cost for autoboxing as in case of `InSet`. As a consequence, we can substantially outperform `InSet` even on 250+ elements.

This PR does not cover long values as Java `switch` statements cannot be used on them. However, we can have a follow-up PR with an approach similar to binary search.

## How was this patch tested?

### Correctness

There is a new test that verifies the logic of the proposed optimization.

### Performance Benchmarks

The performance impact was measured by the benchmarks below (an appropriate part of them will be included in [SPARK-26203](https://issues.apache.org/jira/browse/SPARK-26203)).

```
  def compactBytesBenchmark(numItems: Int, numRows: Long, minNumIters: Int): Benchmark = {
    val name = s"$numItems compact bytes"
    val values = (1 to numItems).map(v => s"CAST($v AS tinyint)")
    val df = spark.range(1, numRows).select($"id".cast(ByteType))
    benchmark(name, df, values, numRows, minNumIters)
  }

  def nonCompactBytesBenchmark(numItems: Int, numRows: Long, minNumIters: Int): Benchmark = {
    val step = (Byte.MaxValue.toInt - Byte.MinValue.toInt) / numItems
    val maxValue = Byte.MinValue + numItems * step
    val rangeSize = maxValue - Byte.MinValue
    require(isLookupSwitch(rangeSize, numItems))
    val name = s"$numItems non-compact bytes"
    val values = (Byte.MinValue until maxValue by step).map(v => s"CAST($v AS tinyint)")
    val df = spark.range(1, numRows).select($"id".cast(ByteType))
    benchmark(name, df, values, numRows, minNumIters)
  }

  def compactShortsBenchmark(numItems: Int, numRows: Long, minNumIters: Int): Benchmark = {
    val name = s"$numItems compact shorts"
    val values = (1 to numItems).map(v => s"CAST($v AS smallint)")
    val df = spark.range(1, numRows).select($"id".cast(ShortType))
    benchmark(name, df, values, numRows, minNumIters)
  }

  def nonCompactShortsBenchmark(numItems: Int, numRows: Long, minNumIters: Int): Benchmark = {
    val step = (Short.MaxValue.toInt - Short.MinValue.toInt) / numItems
    val maxValue = Short.MinValue + numItems * step
    val rangeSize = maxValue - Short.MinValue
    require(isLookupSwitch(rangeSize, numItems))
    val name = s"$numItems non-compact shorts"
    val values = (Short.MinValue until maxValue by step).map(v => s"CAST($v AS smallint)")
    val df = spark.range(1, numRows).select($"id".cast(ShortType))
    benchmark(name, df, values, numRows, minNumIters)
  }

  def compactIntsBenchmark(numItems: Int, numRows: Long, minNumIters: Int): Benchmark = {
    val name = s"$numItems compact ints"
    val values = (1 to numItems).map(v => 10000000 + v)
    val df = spark.range(1, numRows).select($"id".cast(IntegerType))
    benchmark(name, df, values, numRows, minNumIters)
  }

  def nonCompactIntsBenchmark(numItems: Int, numRows: Long, minNumIters: Int): Benchmark = {
    val step = (Int.MaxValue.toLong - Int.MinValue.toLong) / numItems
    val maxValue = Int.MinValue + numItems * step
    val rangeSize = maxValue - Int.MinValue
    require(isLookupSwitch(rangeSize, numItems))
    val name = s"$numItems non-compact ints"
    val values = Int.MinValue until maxValue.toInt by step.toInt
    val df = spark.range(1, numRows).select($"id".cast(IntegerType))
    benchmark(name, df, values, numRows, minNumIters)
  }

  def benchmark(
      name: String,
      df: DataFrame,
      values: Seq[Any],
      numRows: Long,
      minNumIters: Int): Benchmark = {

    val benchmark = new Benchmark(name, numRows, minNumIters, output = output)

    df.createOrReplaceTempView("t")

    val testClosure = () => {
      val df = spark.sql(s"SELECT * FROM t WHERE id IN (${values mkString ","})")
      df.queryExecution.toRdd.foreach(_ => Unit)
    }

    benchmark.addCase("In expression") { _ =>
      withSQLConf(SQLConf.OPTIMIZER_INSET_CONVERSION_THRESHOLD.key -> values.size.toString) {
        testClosure()
      }
    }

    benchmark.addCase("InSet expression") { _ =>
      withSQLConf(SQLConf.OPTIMIZER_INSET_CONVERSION_THRESHOLD.key -> "1") {
        testClosure()
      }
    }

    benchmark
  }

  // this logic is derived from com.sun.tools.javac.jvm.Gen.java
  private def isLookupSwitch(rangeSize: Long, numLabels: Int): Boolean = {
    val tableSpaceCost = 4 + rangeSize
    val tableTimeCost = 3
    val lookupSpaceCost = 3 + 2 * numLabels
    val lookupTimeCost = numLabels
    lookupSpaceCost + 3 * lookupTimeCost < tableSpaceCost + 3 * tableTimeCost
  }

```

The actual benchmarking code looks like:

```
val numRows = 100000000
val minNumIters = 10

compactBytesBenchmark(numItems = 10, numRows = numRows, minNumIters).run()
nonCompactBytesBenchmark(numItems = 10, numRows = numRows, minNumIters).run()
compactBytesBenchmark(numItems = 25, numRows = numRows, minNumIters).run()
nonCompactBytesBenchmark(numItems = 25, numRows = numRows, minNumIters).run()

compactShortsBenchmark(numItems = 10, numRows = numRows, minNumIters).run()
nonCompactShortsBenchmark(numItems = 10, numRows = numRows, minNumIters).run()
compactShortsBenchmark(numItems = 50, numRows = numRows, minNumIters).run()
nonCompactShortsBenchmark(numItems = 50, numRows = numRows, minNumIters).run()
compactShortsBenchmark(numItems = 100, numRows = numRows, minNumIters).run()
nonCompactShortsBenchmark(numItems = 100, numRows = numRows, minNumIters).run()

compactIntsBenchmark(numItems = 10, numRows = numRows, minNumIters).run()
nonCompactIntsBenchmark(numItems = 10, numRows = numRows, minNumIters).run()
compactIntsBenchmark(numItems = 100, numRows = numRows, minNumIters).run()
nonCompactIntsBenchmark(numItems = 100, numRows = numRows, minNumIters).run()
compactIntsBenchmark(numItems = 250, numRows = numRows, minNumIters).run()
nonCompactIntsBenchmark(numItems = 250, numRows = numRows, minNumIters).run()
```

#### Byte

**_Performance Summary_**

Proposed Solution

|  # values       | InSet           | In with compact data | In with non-compact data |
| ------------- |:-------------:| -----:| ---:|
| 10      | 230 ms | 73 ms | 87 ms |
| 25      | 290 ms | 110 ms | 117 ms |

Current Solution

|  # values       | InSet           | In with compact data | In with non-compact data |
| ------------- |:-------------:| -----:| ---:|
| 10      | 220 ms | 91 ms | 91 ms |
| 25      | 291 ms | 167 ms | 168 ms |

**_Raw Results_**

Proposed Solution

```
Running benchmark: 10 compact bytes
  Running case: In expression
  Stopped after 28 iterations, 2053 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3277 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 compact bytes:                        Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   68 /   73       1474.6           0.7       1.0X
InSet expression                               216 /  218        463.9           2.2       0.3X

Running benchmark: 10 non-compact bytes
  Running case: In expression
  Stopped after 24 iterations, 2076 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3340 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 non-compact bytes:                    Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   84 /   87       1186.6           0.8       1.0X
InSet expression                               217 /  223        460.7           2.2       0.4X

Running benchmark: 25 compact bytes
  Running case: In expression
  Stopped after 19 iterations, 2080 ms
  Running case: InSet expression
  Stopped after 15 iterations, 4285 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
25 compact bytes:                        Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                  105 /  110        955.0           1.0       1.0X
InSet expression                               273 /  286        365.8           2.7       0.4X

Running benchmark: 25 non-compact bytes
  Running case: In expression
  Stopped after 18 iterations, 2104 ms
  Running case: InSet expression
  Stopped after 15 iterations, 4449 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
25 non-compact bytes:                    Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                  107 /  117        938.5           1.1       1.0X
InSet expression                               284 /  297        352.3           2.8       0.4X
```

Current Solution

```
Running benchmark: 10 compact bytes
  Running case: In expression
  Stopped after 22 iterations, 2002 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3601 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 compact bytes:                        Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   86 /   91       1168.9           0.9       1.0X
InSet expression                               226 /  240        442.5           2.3       0.4X

Running benchmark: 10 non-compact bytes
  Running case: In expression
  Stopped after 23 iterations, 2085 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3265 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 non-compact bytes:                    Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   86 /   91       1169.1           0.9       1.0X
InSet expression                               213 /  218        469.5           2.1       0.4X

Running benchmark: 25 compact bytes
  Running case: In expression
  Stopped after 15 iterations, 2499 ms
  Running case: InSet expression
  Stopped after 15 iterations, 4326 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
25 compact bytes:                        Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                  164 /  167        610.7           1.6       1.0X
InSet expression                               282 /  288        354.2           2.8       0.6X

Running benchmark: 25 non-compact bytes
  Running case: In expression
  Stopped after 15 iterations, 2518 ms
  Running case: InSet expression
  Stopped after 15 iterations, 4404 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
25 non-compact bytes:                    Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                  164 /  168        611.5           1.6       1.0X
InSet expression                               282 /  294        354.4           2.8       0.6X
```

**_Conclusion_**

As expected, we see that the proposed logic improves the performance on compact and non-compact sequences of byte values. While the performance on compact elements is better than on non-compact ones, we do not see that it is stable independently of the number of elements inside `In`. Hence, we cannot conclude O(1) time performance on compact byte values. Spark generates `(byte) 1`, `(byte) 100` to represent byte literals, so it can potentially be a reason for the non-linear time complexity.

We also see that the performance of the old approach on bytes is the same independently of data compactness (as expected).

#### Short

**_Performance Summary_**

Proposed Solution

|  # values       | InSet           | In with compact data | In with non-compact data |
| ------------- |:-------------:| -----:| ---:|
| 10      | 243 ms | 45 ms | 56 ms |
| 50      | 294 ms | 45 ms | 71 ms |
| 100      | 260 ms | 46 ms | 82 ms |

Current Solution

|  # values       | InSet           | In with compact data | In with non-compact data |
| ------------- |:-------------:| -----:| ---:|
| 10      | 248 ms | 71 ms | 69 ms |
| 50      | 295 ms | 169 ms | 173 ms |
| 100      | 265 ms | 335 ms | 370 ms |

**_Raw Results_**

Proposed Solution

```
Running benchmark: 10 compact shorts
  Running case: In expression
  Stopped after 45 iterations, 2032 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3633 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 compact shorts:                       Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   41 /   45       2430.9           0.4       1.0X
InSet expression                               232 /  242        430.4           2.3       0.2X

Running benchmark: 10 non-compact shorts
  Running case: In expression
  Stopped after 36 iterations, 2012 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3645 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 non-compact shorts:                   Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   53 /   56       1880.2           0.5       1.0X
InSet expression                               233 /  243        428.4           2.3       0.2X

Running benchmark: 50 compact shorts
  Running case: In expression
  Stopped after 45 iterations, 2007 ms
  Running case: InSet expression
  Stopped after 15 iterations, 4366 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
50 compact shorts:                       Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   42 /   45       2368.9           0.4       1.0X
InSet expression                               280 /  291        357.4           2.8       0.2X

Running benchmark: 50 non-compact shorts
  Running case: In expression
  Stopped after 29 iterations, 2053 ms
  Running case: InSet expression
  Stopped after 15 iterations, 4447 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
50 non-compact shorts:                   Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   69 /   71       1451.0           0.7       1.0X
InSet expression                               292 /  297        342.3           2.9       0.2X

Running benchmark: 100 compact shorts
  Running case: In expression
  Stopped after 44 iterations, 2032 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3967 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
100 compact shorts:                      Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   44 /   46       2248.6           0.4       1.0X
InSet expression                               259 /  264        385.6           2.6       0.2X

Running benchmark: 100 non-compact shorts
  Running case: In expression
  Stopped after 25 iterations, 2060 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3832 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
100 non-compact shorts:                  Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   79 /   82       1267.6           0.8       1.0X
InSet expression                               250 /  256        399.4           2.5       0.3X
```
Current Solution

```
Running benchmark: 10 compact shorts
  Running case: In expression
  Stopped after 29 iterations, 2051 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3788 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 compact shorts:                       Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
In expression                                   64 /   71       1553.8           0.6       1.0X
InSet expression                               243 /  253        411.4           2.4       0.3X

Running benchmark: 10 non-compact shorts
  Running case: In expression
  Stopped after 30 iterations, 2055 ms
  Running case: InSet expression
  Stopped after 15 iterations, 3658 ms

Java HotSpot(TM) 64-Bit Server VM 1.8.0_181-b13 on Mac OS X 10.14
Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
10 non-compact shorts:                   Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
…