@dbtsai 2018-11-28
The approach looks great, and can significantly improve the performance. For Long, I agree that we should also implement binary search approach for `O(logn)` look up. 

Wondering which one will be faster, binary search using arrays or rewrite the `if-else` in binary search form.