@aokolnychyi 2018-11-29
@cloud-fan, yeah, let’s see if this PR is useful.

The original idea wasn’t to avoid fixing autoboxing in `InSet`. `In` was tested on 250 numbers to prove O(1) time complexity on compact values and outline problems with `InSet`. After this change, `In` will be faster than `InSet` but this is not the goal. Overall, the intent was to have a tiny and straightforward change that would optimize `In` expressions even if the number of elements is less than “spark.sql.optimizer.inSetConversionThreshold” and Spark does not use `InSet`.

Once we solve autoboxing issues in `InSet`, we would need to benchmark against this approach.